# [DBAL-1106] Improve schema introspection performance on Oracle