@Ocramius 2015-01-04
Looks good, but will only land in `2.6` due to the risk of regressions